store/requirements.txt
fastapi[all]
python-dotenv
alembic
sqlalchemy
psycopg2-binary